requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
decorator==5.1.1
deprecation==2.1.0
idna==3.7
importlib-metadata==6.11.0
mashumaro==3.13.1
packaging==24.1
py==1.11.0
pysnyk==0.9.19
requests==2.32.3
retry==0.9.2
typing_extensions==4.12.2
urllib3==2.2.2
XlsxWriter==3.2.0
zipp==3.19.2
